Encapsulate remote logging into its own package
The logging code got a little unwieldy with it having to support multiple services and stuff. Moving it into its own package so we don't have such a cluttered namespace for types and names. Reduces the surface area of the exported types greatly too.
